requirements.txt
selenium
pytest
webdriver_manager
pyyaml
allure-pytest